+++ src/@types/authorization.d.ts
@@ -1,14 +1,5 @@
-export type AuthorizationTokens = {
-  access_token: string;
-  authuser: number;
-  expires_in: number;
-  prompt: "consent";
-  scope: string;
-  token_type: "Bearer";
-  state: string;
-};
-
-export type AuthorizationError = {
-  error: string;
-  state: string;
+export type AuthorizationService = {
+  id: string;
+  name: string;
+  icon: string;
 };
