+++ src/@types/legacy-authorization.d.ts
@@ -1,0 +1,14 @@
+export type AuthorizationTokens = {
+  access_token: string;
+  authuser: number;
+  expires_in: number;
+  prompt: "consent";
+  scope: string;
+  token_type: "Bearer";
+  state: string;
+};
+
+export type AuthorizationError = {
+  error: string;
+  state: string;
+};
