+++ src/components/login/services.ts
@@ -1,0 +1,13 @@
+//todo: バックエンドから取得するようにする
+
+import { AuthorizationService } from "@/@types/authorization";
+
+const services: AuthorizationService[] = [
+  {
+    id: "google",
+    name: "Google",
+    icon: "/images/google.svg",
+  },
+];
+
+export { services };
